fix a minor miscalculation

+++ software/color.c
@@ -3,7 +3,7 @@
 // HSV to RGB color space conversion.
 // We use a HSV range from 0 .. 1536 so every (3*8bit) RGB value can be expressed as a HSV value.
 void color_hsv2rgb(struct color_hsv* in, struct color_rgb* out) {
-  uint8_t col = (in->v * in->s) / 255; // Span between darkest and brightest channel.
+  uint8_t col = (uint16_t)(in->v * in->s) / 255; // Span between darkest and brightest channel.
   uint8_t min = in->v - col; // Minimum (brightness of the darkest channel).
   uint8_t ramp = in->h >> 8; // Ramp of the HSV diagram we're on.
   uint8_t hue = in->h % 256; // Position on hue ramp.
